# internal-server/swagger.json
{
  "info": {
    "title": "fusion-swagger",
    "version": "1.0.0",
    "description": "internal-server swagger"
  },
  "openapi": "3.0.0",
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "authorization"
      }
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/articles/": {
      "get": {
        "summary": "GET all articles",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "filter",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string"
                }
              },
              "example": {
                "title": "string"
              }
            },
            "description": "filter for where"
          }
        ],
        "tags": [
          "articles"
        ],
        "responses": {
          "200": {
            "description": "articles"
          },
          "500": {
            "description": "Error message. Probably invalid request data"
          }
        }
      },
      "post": {
        "summary": "Create new article",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/createArticleModel"
              }
            },
            "description": "New article object"
          }
        },
        "tags": [
          "articles"
        ],
        "responses": {
          "200": {
            "description": "Article created"
          },
          "424": {
            "description": "Invalid URL"
          },
          "500": {
            "description": "Error message. Probably invalid request data"
          }
        }
      }
    },
    "/api/articles/{id}": {
      "delete": {
        "summary": "Delete article",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "article db PK",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "tags": [
          "articles"
        ],
        "responses": {
          "200": {
            "description": "article deleted"
          },
          "500": {
            "description": "Error message. Probably invalid request data"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "createArticleModel": {
      "type": "object",
      "properties": {
        "link": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "collectionFormat": "multi",
          "items": {
            "type": "object",
            "properties": {
              "value": {
                "type": "string"
              },
              "label": {
                "type": "string"
              },
              "__isNew__": {
                "type": "boolean"
              }
            }
          }
        }
      },
      "required": [
        "link"
      ]
    },
...
  }
}